refactor([issue-1152]): decompose arcPlanner.js into per-pass modules behind a barrel
arcPlanner.js was 2137 lines mixing four LLM-call-chain concerns plus a
shared context layer. Split into ./arcPlanner/:
- context.js       — error helpers, world-context loading, manuscript/issue
                     collectors, prompt-context builders, verify/resolve shapers
- arcCore.js       — the overview ↔ reader-map ↔ refine ↔ verify ↔ resolve ↔
                     commit cluster (mutually recursive, kept together)
- episodeSeedPass.js — season→episode seeding + commit-to-issues + canon extract
- completenessPass.js — manuscript-completeness analysis (+ optional edits)
- manuscriptDerive.js — derive an arc from source text and commit it
- coverConcepts.js  — volume + comic cover-concept generation

arcPlanner.js stays as a re-exporting barrel so every existing import
survives; the __testing internals bundle is reassembled from the new homes.
Extraction was byte-faithful (verified) and the module layering is acyclic
(context ← everything; manuscriptDerive also ← arcCore). New
arcPlannerBarrel.test.js pins same-object re-export, the public entry points,
the __testing bundle, and cross-module name-collision freedom.

Refs #1152
